dtype

+++ src/yahtzee_agent/trainer.py
@@ -442,7 +442,9 @@ class YahtzeeAgentTrainer(lightning.LightningModule):
             )
             self.log("train/bonus_likelihood_loss_raw", bonus_likelihood_loss, prog_bar=False)
             self.log("train/pred_upper", bonus_likelihood_logit.mean().item() * 63, prog_bar=False)
-            self.log("train/target_upper", received_bonus.mean().item() * 63, prog_bar=False)
+            self.log(
+                "train/target_upper", received_bonus.float().mean().item() * 63, prog_bar=False
+            )
 
         self.log("train/total_loss", loss, prog_bar=True)
         self.log("lr", self.trainer.optimizers[0].param_groups[0]["lr"], prog_bar=False)
